// packages/core/src/core/EventDispatcher.d.ts
export interface KeyboardEventListener {
    (event: KeyboardEvent): void;
}
export interface PointerEventListener {
    (event: PointerEvent): void;
}
export interface CustomEventListener {
    (event: CustomEvent): void;
}
export interface FocusEventListener {
    (event: FocusEvent): void;
}
export interface TouchEventListener {
    (event: TouchEvent): void;
}
export interface ChangeEventListener {
    (event: ChangeEvent): void;
}
export interface IoEventListener {
    (event: {
        detail: any;
        target: ReactiveNode | IoElement | EventTarget;
        path: Array<ReactiveNode | IoElement | EventTarget>;
    }): void;
}
export type AnyEventListener = EventListener | KeyboardEventListener | PointerEventListener | CustomEventListener | FocusEventListener | TouchEventListener | ChangeEventListener | IoEventListener | EventListenerOrEventListenerObject;
/**
 * Listener definition type.
 * The first item is a string (function name) or an event listener function.
 * The second item is an optional object of event listener options.
 */
export type ListenerDefinition = [string | AnyEventListener, AddEventListenerOptions?];
/**
 * Loose listener definition type.
 * It can be a string (function name), an event AnyEventListener function or a ListenerDefinition array.
 */
export type ListenerDefinitionLoose = string | AnyEventListener | ListenerDefinition;
export type Listener = [AnyEventListener, AddEventListenerOptions?];
export type Listeners = Record<string, Listener[]>;
// ...
export declare class EventDispatcher {
    readonly node: ReactiveNode | IoElement | EventTarget;
    readonly nodeIsEventTarget: boolean;
    readonly protoListeners: Listeners;
    readonly propListeners: Listeners;
    readonly addedListeners: Listeners;
    /**
     * Creates an instance of `EventDispatcher` for specified `ReactiveNode` instance.
     * It initializes `protoListeners` from `ProtoChain`.
     * @param {ReactiveNode | IoElement | EventTarget} node owner ReactiveNode
     */
    constructor(node: ReactiveNode | IoElement | EventTarget);
    /**
     * Sets `protoListeners` specified as `get Listeners()` class definitions.
     * Definitions from subclass replace the ones from parent class.
     * @param {ReactiveNode | IoElement} node owner ReactiveNode
     */
    setProtoListeners(node: ReactiveNode | IoElement): void;
    /**
     * Sets `propListeners` specified as inline properties prefixed with "@".
     * It removes existing `propListeners` that are no longer specified and it replaces the ones that changed.
     * @param {Record<string, any>} properties - Inline properties
     */
    applyPropListeners(properties: Record<string, any>): void;
    /**
     * Proxy for `addEventListener` method.
     * Adds an event listener to the node's `addedListeners` collection.
     * If the node is an EventTarget, also registers the listener with the DOM.
     * @param {string} name - Name of the event
     * @param {AnyEventListener} listener - Event listener handler
     * @param {AddEventListenerOptions} [options] - Event listener options
     */
    addEventListener(name: string, listener: AnyEventListener, options?: AddEventListenerOptions): void;
    /**
     * Proxy for `removeEventListener` method.
     * Removes an event listener from the node's `addedListeners` collection.
     * If `listener` is not specified it removes all listeners for specified `type`.
     * @param {string} name - Name of the event
     * @param {AnyEventListener} listener - Event listener handler
     * @param {AddEventListenerOptions} [options] - Event listener options
    */
    removeEventListener(name: string, listener?: AnyEventListener, options?: AddEventListenerOptions): void;
    /**
     * Shorthand for custom event dispatch.
     * @param {string} name - Name of the event
     * @param {any} detail - Event detail data
     * @param {boolean} [bubbles] - Makes event bubble
     * @param {ReactiveNode | IoElement | EventTarget} [node] - Event target override to dispatch the event from
     */
    dispatchEvent(name: string, detail?: any, bubbles?: boolean, node?: ReactiveNode | IoElement | EventTarget, path?: Array<ReactiveNode | IoElement | EventTarget>, visited?: Set<ReactiveNode | IoElement | EventTarget>): void;
    /**
     * Disconnects all event listeners and removes all references for garbage collection.
     * Use this when node is discarded.
     */
    dispose(): void;
}